Create Floating point comparison (floats have limited precision and are unable to exactly represent some values. Rounding errors accumulate with repeated computation, and numbers expected to be equal often differ slightly.  As a result, it is common advice to not use an exact equality comparison (==) with floats.)
floats have limited precision and are unable to exactly represent some values. Rounding errors accumulate with repeated computation, and numbers expected to be equal often differ slightly.

As a result, it is common advice to not use an exact equality comparison (==) with floats.

Task
You have:

a float value that comes from a computation and may have accumulated errors up to ±0.001
a reference value
a function approx_equals that compare the two values taking into account loss of precision; the function should return true if and only if the two values are close to each other, the maximum allowed difference is 0.001
The function is bugged and sometimes returns wrong results.

Your task is to correct the bug.